Somehow we missed this bugfix from Mihai. However, the actual patch has to be a little different from Mihai's, as cont is an attribute from options:
My original patch introduced a fairly serious bug, we need to apply this bugfix as soon as possible.

On Fri, 2009-05-29 at 18:55 -0700, Mihai Rusu wrote:
> On Monday 04 May 2009 13:42:18 Lucas Meneghel Rodrigues wrote:
> > Summary: If autotest is not running with the --continue flag,
> > delete the state file transparently
> >
> > For quite a while, people developing tests had problems when they
> > had to interrupt a given test before it ends: it's not possible to
> > guarantee the consistency of the state file, hence the next time
> > the test is executed, an exception is thrown. The workaround is, of
> > course, remove the state file allways. However, as Martin pointed
> > out, when autotest is not running with the flag -c or --continue,
> > we can just remove the previously left state file to make this
> > process transparent to the test developer.
> >
> >
> Unfortunately the patch misses any check if continuation has been 
> requested thus current SVN is unable to execute any jobs that require 
> continuation (for example kernel tests that reboot the machine). 
> Please apply the following patch to fix it:
> 
>  --- autotest/client/bin/job.py
> +++ autotest/client/bin/job.py
> @@ -1251,7 +1251,7 @@
>      state = control + '.state'
>      # Ensure state file is cleaned up before the job starts to run if 
> autotest
>      # is not running with the --continue flag
> -    if os.path.isfile(state):
> +    if not cont and os.path.isfile(state):
>          logging.debug('Cleaning up previously found state file')
>          os.remove(state)

diff --git a/client/bin/job.py b/client/bin/job.py
index 437b729..6f8f650 100755

+++ client/bin/job.py
@@ -1288,7 +1288,7 @@ def runjob(control, options):
     state = control + '.state'
     # Ensure state file is cleaned up before the job starts to run if autotest
     # is not running with the --continue flag
-    if os.path.isfile(state):
+    if not options.cont and os.path.isfile(state):
         logging.debug('Cleaning up previously found state file')
         os.remove(state)
